pyproject.toml
[tool.poetry]
name = "modos"
version = "0.1.0"
description = "SMOC Multi-Omics Digital Object System API"
authors = ["SDSC-ORDES"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "modos" }]

[tool.poetry.dependencies]
python = "^3.10"
rdflib = "^6.3"
zarr = "^2.16.1"
calamus = "^0.4.2"
pysam = "^0.22.0"
modos_schema = { git = "https://github.com/sdsc-ordes/modos-schema.git", branch = "main" }
typer = "^0.9.0"
s3fs = "^2024.3.1"
htsget = "^0.2.6"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.0"
pytest-cov = "^4.1.0"
testcontainers = { extras = ["minio"], version = "^4.4.1" }

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
myst-parser = "^2.0.0"
sphinx = "^7.2.6"
pydata-sphinx-theme = "^0.15.2"
sphinx-click = "^5.1.0"
sphinx-autoapi = "^3.0.0"
sphinx-design = "^0.5.0"

[tool.poetry.scripts]
modos = 'modos.cli:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Tooling configuration

[tool.pytest.ini_options]
addopts = ["--doctest-modules"]
testpaths = ["modos", "tests"]

[tool.black]
line-length = 79
target-version = ["py310"]
